fix: callForRawType calls with 32 bits instead of 64 bits

+++ include/SupportedUserTypes.h
@@ -977,7 +977,7 @@ namespace ChimeraTK {
         lambda(int32_t());
       } break;
       case DataType::int64: {
-        lambda(int32_t());
+        lambda(int64_t());
       } break;
       // NOLINTEND(bugprone-branch-clone)
       default:
